Improvement  in notifications

+++ database/migrations/2021_11_25_201155_add_cascade_on_delete_to_invoice_and_invoice_entries.php
@@ -1,8 +1,9 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 class AddCascadeOnDeleteToInvoiceAndInvoiceEntries extends Migration
 {
@@ -14,17 +15,25 @@ class AddCascadeOnDeleteToInvoiceAndInvoiceEntries extends Migration
     public function up()
     {
         Schema::table('invoices', function (Blueprint $table) {
-            $table->dropForeign('invoices_card_id_foreign');
-            $table->foreign('card_id')
-                ->references('id')->on('cards')
-                ->onDelete('cascade');
+            if (env('DB_CONNECTION') !== 'sqlite') {
+                $table->dropForeign('invoices_card_id_foreign');
+                
+                $table->foreign('card_id')
+                    ->references('id')->on('cards')
+                    ->onDelete('cascade');
+            }
+
         });
 
         Schema::table('invoice_entries', function (Blueprint $table) {
-            $table->dropForeign('invoice_entries_invoice_id_foreign');
-            $table->foreign('invoice_id')
-                ->references('id')->on('invoices')
-                ->onDelete('cascade');
+            if (env('DB_CONNECTION') !== 'sqlite') {
+                $table->dropForeign('invoice_entries_invoice_id_foreign');
+
+                $table->foreign('invoice_id')
+                    ->references('id')->on('invoices')
+                    ->onDelete('cascade');
+            }
+
         });
     }
